:memo: 增加部分功能，考勤管理，优化部分逻辑处理。增加员工端处理。包括员工打卡，员工信息，个人中心，密码修改，个人奖惩等。员工处理为在hr表新增字段映射员工表，统计相关信息，利用hr表登陆。

+++ src/main/java/com/liruilong/hros/service/DepartmentService.java
@@ -13,7 +13,7 @@ import java.util.Date;
 import java.util.List;
 
 /**
- * @Author Liruilong
+ * @Author liruilong
  * @Description
  * @Date 23:29 2019/12/28
  * @Param
@@ -40,8 +40,13 @@ public class DepartmentService {
     }
 
     public void deleteDepById(Department dep) {
+        List<Department> withOutChildren = departmentMapper.getAllDepartmentsWithOutChildren();
+        if (withOutChildren.size() <= 1) {
+            dep.setResult(-3);
+            return;
+        }
         departmentMapper.deleteDepById(dep);
-        oplogService.addOpLog(new OpLog((byte) 1,new Date(),"删除部门:id=" + dep.getId(), Hruitls.getCurrent().getName()));
+        oplogService.addOpLog(new OpLog((byte) 1, new Date(), "删除部门:id=" + dep.getId(), Hruitls.getCurrent().getName()));
     }
